moving gui functions to gui.go
